backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic-settings==2.1.0
supabase==2.3.0
gotrue==2.4.4
python-dotenv==1.0.0
httpx==0.24.0
numpy==1.26.2
APScheduler==3.10.4
pytz==2023.3